Move shared_tasks.py into shared_tasks directory
