+++ dags/dev_validate_by_mapper_type_to_gdrive.py
@@ -11,7 +11,7 @@ from rikolti.dags.utils_by_mapper_type import make_mapper_type_endpoint
 from rikolti.dags.utils_by_mapper_type import fetch_endpoint_task
 from rikolti.dags.utils_by_mapper_type import map_endpoint_task
 from rikolti.dags.utils_by_mapper_type import validate_endpoint_task
-from rikolti.dags.shared_tasks import s3_to_localfilesystem
+from rikolti.dags.shared_tasks.shared_tasks import s3_to_localfilesystem
 
 logger = logging.getLogger("airflow.task")
